// compiler/src/symbol-table/symbol-table.interface.d.ts
export type FPLDataType = 'int' | 'decimal' | 'currency' | 'percentage' | 'boolean' | 'string' | 'date' | 'array' | 'object' | 'customer' | 'loan' | 'account' | 'policy_result' | 'policy' | 'function' | 'void' | 'null' | 'unknown';